[fix][Step2][피드백 반영] 패키지 수정 step2 -> lotto
