+++ src/main/java/lotto/InputValidator.java
@@ -1,4 +1,4 @@
-package step2;
+package lotto;
 
 public class InputValidator {
