+++ src/main/java/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package step2;
+package lotto;
 
 import java.util.ArrayList;
 import java.util.Collections;
