+++ src/main/java/lotto/LottoTicket.java
@@ -1,4 +1,4 @@
-package step2;
+package lotto;
 
 import java.util.Arrays;
 import java.util.List;
